authentication added

+++ server/schemas/resolvers.js
@@ -16,9 +16,14 @@ const resolvers = {
 
 			throw new AuthenticationError('Not logged in');
 		},
-		getTodos: async () => {
-			const todos = await Todo.find();
-			return todos;
+		getTodos: async (parent, args, context) => {
+			if (context.user) {
+				const todos = await Todo.find();
+				return todos;
+			}
+
+			throw new AuthenticationError('Please Login');
+
 		},
 		getTodo: async (root, args) => {
 			const todo = await Todo.findById(args.id);
@@ -37,51 +42,62 @@ const resolvers = {
 
 		login: async (parent, { email, password }) => {
 			const user = await User.findOne({ email });
-	  
+
 			if (!user) {
-			  throw new AuthenticationError('No User with this email found!');
+				throw new AuthenticationError('No User with this email found!');
 			}
-	  
+
 			const correctPw = await user.isCorrectPassword(password);
-	  
+
 			if (!correctPw) {
-			  throw new AuthenticationError('Incorrect password!');
+				throw new AuthenticationError('Incorrect password!');
 			}
-	  
+
 			const token = signToken(user);
 			return { token, user };
-		  },
+		},
 
 
-		addTodo: async (root, args) => {
-			const newTodo = new Todo({ title: args.title, detail: args.detail, date: args.date })
-			await newTodo.save()
-			return newTodo
+		addTodo: async (root, args, context) => {
+			if (context.user) {
+				const newTodo = new Todo({ title: args.title, detail: args.detail, date: args.date })
+				await newTodo.save()
+				return newTodo
+			}
+			throw new AuthenticationError('Please Login to Add');
+
 		},
-		deleteTodo: async (root, args) => {
-			await Todo.findByIdAndDelete(args.id);
-            return 'Your todo has been deleted sucessfully'
-        },
-			
+		deleteTodo: async (root, args, context) => {
+			if (context.user) {
+				await Todo.findByIdAndDelete(args.id);
+				return 'Your todo has been deleted sucessfully'
+			}
+			throw new AuthenticationError('Please Login to Delete');
 		},
-		updateTodo:async (root, args) => {
-			const{id,title,detail,date} = args
-			const updatedTodo = {};
-			
-			if(title!=undefined){
-				updatedTodo.title = title;
+		updateTodo: async (root, args, context) => {
+			if (context.user) {
+				const { id, title, detail, date } = args
+				const updatedTodo = {};
+
+				if (title != undefined) {
+					updatedTodo.title = title;
+				}
+				if (detail != undefined) {
+					updatedTodo.detail = detail;
+				}
+				if (date != undefined) {
+					updatedTodo.date = date;
+				}
+				const todo = await Todo.findByIdAndUpdate(id, updatedTodo, { new: true })
+				return todo;
 			}
-			if(detail!=undefined){
-				updatedTodo.detail = detail;
-			}
-			if(date!=undefined){
-				updatedTodo.date = date;
-			}
-			const todo = await Todo.findByIdAndUpdate(id,updatedTodo,{new:true})
-			return todo;
+			throw new AuthenticationError('Please Login to Update');
 		}
-	}
+
+	},
+
 }
+
 
 
 module.exports = resolvers;
